requirements.txt
# Core Framework
fastapi==0.118.2
uvicorn==0.37.0
pydantic==2.12.0
pydantic-settings==2.11.0

# Machine Learning
scikit-surprise==1.1.4
pandas==2.3.3
numpy==1.26.4
scikit-learn==1.7.2

# MLOps
mlflow==3.4.0

# Database
psycopg2-binary==2.9.10
sqlalchemy==2.0.43

# Cache
redis==6.4.0

# HTTP Client
httpx==0.28.1
requests==2.32.5

# Data Processing
python-dotenv==1.1.1
pyyaml==6.0.3

# Utilities
python-dateutil==2.9.0.post0
pytz==2025.2

# Development & Testing
pytest==8.4.2
pytest-asyncio==1.2.0
black==25.9.0
flake8==7.3.0

# Logging
python-json-logger==4.0.0

# dummy data
faker==37.11.0